Add event management service and DTOs with controller integration
Introduce `EventApplicationService` for event operations including create, publish, cancel, and registration management. Add DTOs such as `CreateEventRequest`, `EventResponse`, and mappers for handling API input/output. Update persistence layer and controller for event creation via REST API.

+++ src/main/java/com/grupocaos/products/athletix/event/infrastructure/persistence/entity/LocationEmbeddable.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * Represents an embeddable value object for storing location-related data.
  * <p>
@@ -30,8 +32,8 @@ public class LocationEmbeddable {
     String country;
 
     @Column(precision = 10, scale = 7)
-    Double latitude;
+    BigDecimal latitude;
 
     @Column(precision = 10, scale = 7)
-    Double longitude;
+    BigDecimal longitude;
 }
